to the browser

+++ teamRepo/hpp/ServerConfig.hpp
@@ -27,7 +27,7 @@ class ServerConfig
 		std::string							serverNames; //vector
 		// std::vector<std::string>			index;
 		// std::string						root;
-		std::map<u_int8_t, std::string>		errorPages;
+		std::map<int, std::string>		errorPages;
 		std::map<std::string, location>		allLocations;
 		size_t								limitBodySize;
